Fix (complement) for array reversed problem
The function to_value should be fixed, otherwise arrays in json would appear
inverted in erlang (the problem wasn't only related to proplists). Besides,
the eunit test showing this problem was missing, and the record associated
with it was using a N-tuple rather than a list. (N-tuple is an invalid
datatype for mochijson2). I fixed the record definition and implemented
the test.