Replace mochijson2 for jiffy and fix support for undefined record members.
Jiffy is a much better performing json parser and its convertion table
is very similar to mochijson2's.

Regarding the fix of undefined record members. The problem was that
undefined members were encoded with a string value of "undefined"
leading to a problem when decoding back. In order to fix this
situation, undefined record members are skipped during encoding.
This is compatible with the decoder behavior where a member would not be
assigned if it's missing in the json structure. A record member assigned to
the atom null is encoded properly and decoded back to the same atom null.