Use default values of fields to determine their types
If a field of a record has a default value and this value is a record
exported in any of the provided modules in the to_rec method, this
record will be used as the type of the field. So, in these cases, the
json object, which represents the value of the field, will be
transformed to a record.